unsplit a pointlessly split file
