+++ enclone_tools/src/lib.rs
@@ -3,5 +3,4 @@
 pub mod copy_for_enclone;
 pub mod feature_barcode_matrix;
 pub mod html;
-pub mod main_build_immcantation_inputs;
 pub mod run_test;
